Move ops to dedicated directory, and fix linting issues.
